+++ apps/person-search-service/src/main/java/no/nav/registre/testnorge/personsearchservice/adapter/utils/AlderUtils.java
@@ -1,0 +1,52 @@
+package no.nav.registre.testnorge.personsearchservice.adapter.utils;
+
+import lombok.experimental.UtilityClass;
+import no.nav.registre.testnorge.personsearchservice.controller.search.PersonSearch;
+import org.apache.lucene.search.join.ScoreMode;
+import org.elasticsearch.index.query.BoolQueryBuilder;
+import org.elasticsearch.index.query.QueryBuilders;
+
+import java.time.LocalDate;
+import java.util.Optional;
+
+import static java.util.Objects.nonNull;
+import static no.nav.registre.testnorge.personsearchservice.adapter.utils.QueryUtils.getBetween;
+
+@UtilityClass
+public class AlderUtils {
+
+    public static void addAlderQueries(BoolQueryBuilder queryBuilder, PersonSearch search){
+        addAlderQuery(queryBuilder, search);
+        addFoedselQuery(queryBuilder, search);
+    }
+
+    private static void addFoedselQuery(BoolQueryBuilder queryBuilder, PersonSearch search) {
+        Optional.ofNullable(search.getFoedsel())
+                .ifPresent(value -> queryFoedselsdato(value.getFom(), value.getTom(), queryBuilder));
+    }
+
+    private static void addAlderQuery(BoolQueryBuilder queryBuilder, PersonSearch search) {
+        Optional.ofNullable(search.getAlder())
+                .ifPresent(value -> queryAlder(value.getFra(), value.getTil(), queryBuilder));
+    }
+
+    private static void queryFoedselsdato(LocalDate fom, LocalDate tom, BoolQueryBuilder queryBuilder) {
+        getBetween(fom, tom, "hentPerson.foedsel.foedselsdato")
+                .ifPresent(rangeQueryBuilder -> queryBuilder.must(QueryBuilders.nestedQuery(
+                                "hentPerson.foedsel",
+                                rangeQueryBuilder,
+                                ScoreMode.Avg
+                        ))
+                );
+    }
+
+    private static void queryAlder(Short fra, Short til, BoolQueryBuilder queryBuilder) {
+        LocalDate now = LocalDate.now();
+
+        LocalDate tom = nonNull(fra) ? now.minusYears(fra).minusMonths(3) : now.minusMonths(3);
+        LocalDate fom = nonNull(til) ? now.minusYears(til).minusYears(1) : null;
+
+        queryFoedselsdato(fom, tom, queryBuilder);
+    }
+
+}
