+++ apps/person-search-service/src/main/java/no/nav/registre/testnorge/personsearchservice/adapter/utils/QueryUtils.java
@@ -4,8 +4,13 @@ import lombok.experimental.UtilityClass;
 import org.apache.lucene.search.join.ScoreMode;
 import org.elasticsearch.index.query.NestedQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
+import org.elasticsearch.index.query.RangeQueryBuilder;
 
+import java.time.LocalDate;
 import java.util.Collection;
+import java.util.Optional;
+
+import static java.util.Objects.nonNull;
 
 @UtilityClass
 public class QueryUtils {
@@ -32,5 +37,21 @@ public class QueryUtils {
                 QueryBuilders.existsQuery(path + "." + name),
                 ScoreMode.Avg
         );
+    }
+
+    public static Optional<RangeQueryBuilder> getBetween(LocalDate fom, LocalDate tom, String field) {
+        if (fom == null && tom == null) {
+            return Optional.empty();
+        }
+        var builder = QueryBuilders.rangeQuery(field);
+
+        if (nonNull(fom)) {
+            builder.gte(fom);
+        }
+
+        if (nonNull(tom)) {
+            builder.lte(tom);
+        }
+        return Optional.of(builder);
     }
 }
